Fixed Type of Ammount to Update the Balance

+++ account.py
@@ -17,9 +17,8 @@ class Account :
     
     # Updates the account balance by adding the given amount
     def updateBalance(self,amount):
-        
         try: #Validate that the input can be converted to an integer 
-            amount = int(amount)  
+            amount = float(amount)  
         except ValueError:
             raise TypeError("Only integers are allowed as amount") 
         self.balance += amount
